произведена модуляция: файлы main.js, data.js, util.js
